Removed unnecessary name.
It just complicates matters, when you look at the lock files in /var/db/. With this change the lock file is correctly named `.puppet_pkgdmg_installed_GPGTools`. Including a name like 'Install.pkg' is also confusing to other people, since you could get the idea that it only runs the Install.pkg inside the downloaded .dmg, which is not the case. The `pkgdmg` puppet provider installs all the .pkgs inside the given .dmg file. It so happens that the installer is a .pkg file and the uninstaller is a .app file, so it works out in the end.